+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - Bugfix: nested categories are now matched up correctly between target and replacement
 - Bugfix: Brassica no longer freezes with rules where the target is entirely optional
 - Wildcard symbols can now be used in the replacement of a rule
+- Brassica now applies sound changes to words in parallel, giving a significant speedup on multi-core machines (though not in a webpage)
 
 ### Code
 
@@ -15,6 +16,8 @@
 - `Wildcard` and `Kleene` no longer have `OneOf 'Target 'Env` constraint
 - `OneOf` type family is no longer used and has been removed
 - `Target` and `Environment` `LexemeType`s have been unified as `Matched`
+- `Brassica.SoundChange.Frontend.Internal.parseTokeniseAndApplyRules` now takes another argument specifying how to map over the parse output,
+    allowing it to be run both on a single core and in parallel depending on the provided function
 
 ## v0.2.0